fix: 將Create Shipment Payloads成獨立物件
